improve %b

+++ main.h
@@ -8,7 +8,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-	int _putchar(char c);
+int _putchar(char c);
 int _printf(const char *format, ...);
 char *_get_specifier(const char *string, int start);
 int _get_specifiers_number(const char *string);
@@ -21,7 +21,7 @@ int print_string(char *s);
 int print_non_printable(char *s);
 int print_number(int n);
 int print_unsigned_int(unsigned int n);
-int print_binary(int64_t n);
+int print_binary(uint32_t n);
 int print_octal(int64_t n);
 int print_base16_lower(int64_t n);
 int print_base16_upper(int64_t n);
